refactor(indexer-api): move NoopStorage impls to respective submodules
