+++ indexer-api/src/domain/storage/block.rs
@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::domain::Block;
-use futures::Stream;
+use crate::domain::{Block, storage::NoopStorage};
+use futures::{Stream, stream};
 use indexer_common::domain::BlockHash;
 use std::{fmt::Debug, num::NonZeroU32};
 
@@ -36,4 +36,31 @@ where
         height: u32,
         batch_size: NonZeroU32,
     ) -> impl Stream<Item = Result<Block, sqlx::Error>> + Send;
+}
+
+#[allow(unused_variables)]
+impl BlockStorage for NoopStorage {
+    #[cfg_attr(coverage, coverage(off))]
+    async fn get_latest_block(&self) -> Result<Option<Block>, sqlx::Error> {
+        unimplemented!()
+    }
+
+    #[cfg_attr(coverage, coverage(off))]
+    async fn get_block_by_hash(&self, hash: BlockHash) -> Result<Option<Block>, sqlx::Error> {
+        unimplemented!()
+    }
+
+    #[cfg_attr(coverage, coverage(off))]
+    async fn get_block_by_height(&self, height: u32) -> Result<Option<Block>, sqlx::Error> {
+        unimplemented!()
+    }
+
+    #[cfg_attr(coverage, coverage(off))]
+    fn get_blocks(
+        &self,
+        height: u32,
+        batch_size: NonZeroU32,
+    ) -> impl Stream<Item = Result<Block, sqlx::Error>> {
+        stream::empty()
+    }
 }
